Add support for project specific configuration.
This allows Uncrustify to find configuration in project settings files,
allowing easier per-project configuration.  Configuration paths also have
two expansions on their names done.  The first is for ${project_dir},
which allows a project configuraiton to specify a relative path for the
configuration file.  The second is use of OS environment variable expansions,
allowing ${HOME} and such to be specified - making it easier to share these
configurations between users or across shared filesystems where paths may
differ on different systems.  (Think virtualization for example.)

While here I've updated the documentation in the sample configuration
to show these details, and to correct English language issues hopefully
making it clearer for some users.  (I'm a native English speaker -- it was
clear that the original author is not.)